ltrc/requirements.txt
Django==2.1.3
gunicorn==19.9.0
psycopg2==2.7.6.1
whitenoise==4.1.1
# django-crispy-forms==1.7.0
phonenumbers==8.10.21
django-phonenumber-field==3.0.1
django-import-export==1.2.0
# django-extensions
ipython[notebook]==7.12.0
jupyter==1.0.0